RFCT: Split constants for readability and maintainabiilty
